+++ src/releases/releaseNotes.ts
@@ -5,8 +5,8 @@ export const releaseNotes: Record<
     date?: Date;
   }
 > = {
-  '3.10.3': {
-    isPro: false,
-    date: new Date('2025-11-26'),
+  '3.11.0': {
+    isPro: true,
+    date: new Date('2025-12-01'),
   },
 };